fix(tap): conform to singer.io spec
