+++ setup.py
@@ -3,14 +3,14 @@ from setuptools import setup
 
 setup(
     name="tap-surveymonkey",
-    version="0.1.0",
+    version="0.1.1",
     description="Singer.io tap for extracting data",
     author="Stitch",
     url="http://singer.io",
     classifiers=["Programming Language :: Python :: 3 :: Only"],
     py_modules=["tap_surveymonkey"],
     install_requires=[
-        "singer-python>=5.0.12",
+        "singer-python==5.6.0",
         "requests",
     ],
     entry_points="""
@@ -18,7 +18,7 @@ setup(
     tap-surveymonkey=tap_surveymonkey:main
     """,
     packages=["tap_surveymonkey"],
-    package_data = {
+    package_data={
         "schemas": ["tap_surveymonkey/schemas/*.json"]
     },
     include_package_data=True,
