+++ tap_surveymonkey/schemas/responses.orig.json
@@ -1,0 +1,186 @@
+{
+    "$schema": "http://json-schema.org/draft-06/schema#",
+    "$ref": "#/definitions/SurveyElement",
+    "definitions": {
+        "SurveyElement": {
+            "type": "object",
+            "additionalProperties": false,
+            "properties": {
+                "total_time": {
+                    "type": "integer"
+                },
+                "href": {
+                    "type": "string"
+                },
+                "custom_variables": {
+                    "type": "object"
+                },
+                "ip_address": {
+                    "type": "string"
+                },
+                "id": {
+                    "type": "string"
+                },
+                "logic_path": {
+                    "type": "object"
+                },
+                "date_modified": {
+                    "type": "string",
+                    "format": "date-time"
+                },
+                "response_status": {
+                    "$ref": "#/definitions/ResponseStatus"
+                },
+                "custom_value": {
+                    "type": "string"
+                },
+                "analyze_url": {
+                    "type": "string"
+                },
+                "pages": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Page"
+                    }
+                },
+                "page_path": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "recipient_id": {
+                    "type": "string"
+                },
+                "collector_id": {
+                    "type": "string"
+                },
+                "date_created": {
+                    "type": "string",
+                    "format": "date-time"
+                },
+                "survey_id": {
+                    "type": "string"
+                },
+                "collection_mode": {
+                    "$ref": "#/definitions/CollectionMode"
+                },
+                "edit_url": {
+                    "type": "string"
+                },
+                "metadata": {
+                    "type": "object"
+                }
+            },
+            "required": [
+                "analyze_url",
+                "collection_mode",
+                "collector_id",
+                "custom_value",
+                "custom_variables",
+                "date_created",
+                "date_modified",
+                "edit_url",
+                "href",
+                "id",
+                "ip_address",
+                "logic_path",
+                "metadata",
+                "page_path",
+                "pages",
+                "recipient_id",
+                "response_status",
+                "survey_id",
+                "total_time"
+            ],
+            "title": "SurveyElement"
+        },
+        "Page": {
+            "type": "object",
+            "additionalProperties": false,
+            "properties": {
+                "id": {
+                    "type": "string"
+                },
+                "questions": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Question"
+                    }
+                }
+            },
+            "required": [
+                "id",
+                "questions"
+            ],
+            "title": "Page"
+        },
+        "Question": {
+            "type": "object",
+            "additionalProperties": false,
+            "properties": {
+                "id": {
+                    "type": "string"
+                },
+                "answers": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Answer"
+                    }
+                },
+                "variable_id": {
+                    "type": "string"
+                }
+            },
+            "required": [
+                "answers",
+                "id"
+            ],
+            "title": "Question"
+        },
+        "Answer": {
+            "type": "object",
+            "additionalProperties": false,
+            "properties": {
+                "choice_id": {
+                    "type": "string"
+                },
+                "row_id": {
+                    "type": "string"
+                },
+                "text": {
+                    "type": "string"
+                },
+                "other_id": {
+                    "type": "string"
+                },
+                "col_id": {
+                    "type": "string"
+                }
+            },
+            "required": [],
+            "title": "Answer"
+        },
+        "CollectionMode": {
+            "type": "string",
+            "enum": [
+                "default",
+                "preview",
+                "data_entry",
+                "survey_preview",
+                "edit"
+            ],
+            "title": "CollectionMode"
+        },
+        "ResponseStatus": {
+            "type": "string",
+            "enum": [
+                "partial",
+                "completed",
+                "disqualified",
+                "overquota"
+            ],
+            "title": "ResponseStatus"
+        }
+    }
+}
